RES-2501: Move nupic.research.frameworks.dendrites to packages
