+++ .circleci/config.yml
@@ -90,7 +90,7 @@ jobs:
           key: pip-cache-v1-{{ arch }}-{{ checksum "nupic.torch.sha" }}-{{ checksum "setup.cfg" }}
       - run: pip show flake8-formatter-junit-xml || pip install --prefix ~/.local flake8-formatter-junit-xml
       - run: pip show nupic.torch || pip install --prefix ~/.local git+https://github.com/numenta/nupic.torch.git#egg=nupic.torch
-      - run: pip install --prefix ~/.local -e ".[dev,imagenet,dendrites,gsc]"
+      - run: pip install --prefix ~/.local -e ".[dev,imagenet,gsc]"
       - save_cache:
           key: pip-cache-v1-{{ arch }}-{{ checksum "nupic.torch.sha" }}-{{ checksum "setup.cfg" }}
           paths:
