+++ environment.yml
@@ -19,12 +19,13 @@ dependencies:
 
       # Install nupic.research in development mode with dev tools extras enabled
       - -e ".[dev]"
+      
+      # Install nupic.research common frameworks
+      - -e packages/dendrites
 
       # Install projects extra dependency
       # See `setup.cfg` for available extras.
-      # - -e ".[vernon]"
       # - -e ".[imagenet]"
       # - -e ".[gsc]"
-      # - -e ".[dendrites]"
       # - -e ".[transformers]"
