+++ nupic/research/frameworks/vernon/mixins/__init__.py
@@ -19,19 +19,15 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-from .centroid_context import *
 from .composite_loss import CompositeLoss
 from .configure_optimizer_param_groups import ConfigureOptimizerParamGroups
 from .constrain_parameters import ConstrainParameters
-from .context_signal import *
 from .cutmix import CutMix, CutMixKnowledgeDistillation
 from .delay_load_checkpoint import *
-from .dendrite_metrics import *
 from .ewc import ElasticWeightConsolidation
 from .export_model import ExportModel
 from .extra_validations_per_epoch import *
 from .gradient_metrics import GradientMetrics
-from .hidden_activations import *
 from .knowledge_distillation import KnowledgeDistillation, KnowledgeDistillationCL
 from .legacy_imagenet_config import LegacyImagenetConfig
 from .load_preprocessed_data import LoadPreprocessedData
@@ -52,7 +48,6 @@ from .prune_low_snr import PruneLowSNRGlobal, PruneLowSNRLayers
 from .reduce_lr_after_task import ReduceLRAfterTask
 from .regularize_loss import RegularizeLoss
 from .report_max_accuracy import ReportMaxAccuracy
-from .representation_overlap import *
 from .rezero_weights import RezeroWeights
 from .save_final_checkpoint import SaveFinalCheckpoint
 from .si import *
@@ -60,5 +55,4 @@ from .step_based_logging import *
 from .torch_profiler import *
 from .track_representation_sparsity import *
 from .update_boost_strength import UpdateBoostStrength
-from .update_dendrite_boost_strength import UpdateDendriteBoostStrength
 from .vary_batch_size import VaryBatchSize
