+++ packages/dendrites/nupic/research/frameworks/dendrites/apply_dendrites_hook.py
