+++ packages/dendrites/nupic/research/frameworks/dendrites/functional/apply_dendrites.py
