+++ packages/dendrites/nupic/research/frameworks/dendrites/mixins/centroid_figure_1b.py
