+++ packages/dendrites/nupic/research/frameworks/dendrites/mixins/context_signal.py
