+++ packages/dendrites/nupic/research/frameworks/dendrites/mixins/hidden_activations.py
