+++ packages/dendrites/nupic/research/frameworks/dendrites/mixins/sp_context.py
