+++ packages/dendrites/nupic/research/frameworks/dendrites/modules/apply_dendrites.py
