+++ packages/dendrites/nupic/research/frameworks/dendrites/modules/dendritic_mlp.py
