+++ projects/dendrites/permutedMNIST/experiments/dendrites_xor_sparsity.py
@@ -40,12 +40,13 @@ from nupic.research.frameworks.dendrites import (
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
 )
+from nupic.research.frameworks.dendrites.mixins import CentroidContext
 from nupic.research.frameworks.pytorch.datasets import PermutedMNIST
 from nupic.research.frameworks.vernon import mixins
 
 
 class CentroidExperiment(mixins.RezeroWeights,
-                         mixins.CentroidContext,
+                         CentroidContext,
                          mixins.PermutedMNISTTaskIndices,
                          DendriteContinualLearningExperiment):
     pass
