+++ projects/dendrites/permutedMNIST/experiments/hyperparameter_search.py
@@ -37,6 +37,7 @@ from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
 )
+from nupic.research.frameworks.dendrites.mixins import CentroidContext
 from nupic.research.frameworks.pytorch.datasets import PermutedMNIST
 from nupic.research.frameworks.vernon import mixins
 
@@ -45,7 +46,7 @@ from nupic.research.frameworks.vernon import mixins
 
 class NbSegmentSearchExperiment(
     mixins.RezeroWeights,
-    mixins.CentroidContext,
+    CentroidContext,
     mixins.PermutedMNISTTaskIndices,
     DendriteContinualLearningExperiment,
 ):
